Convert the DAA test to an application.
