+++ stability-tests/daa/config.go
@@ -1,4 +1,4 @@
-package daa
+package main
 
 import (
 	"path/filepath"
